+++ frontend/src/views/OutputHeader.svelte
@@ -7,6 +7,8 @@
   export let inflight = false;
   export let client_stream = false;
   export let server_stream = false;
+  export let outCount = 0;
+  export let inCount = 0;
 
   const onCancelClicked = () => backend.api.Cancel()
   const onCloseClicked = () => backend.api.CloseSend()
@@ -25,7 +27,7 @@
 
 <div class="output-header">
   {#if inflight}
-    <RequestType {client_stream} {server_stream} />
+    <RequestType {outCount} {inCount} {client_stream} {server_stream} />
   {/if}
   {#if !inflight && rpc.status }
     <Status status={rpc.status} code={rpc.status_code} />